refactor: move NamedEventSource to same package as EventSource
